requirements.txt
beautifulsoup4==4.10.0
cloudinary==1.29.0
Pillow==8.4.0
pydantic==2.0.3
pydantic_core==2.3.0
PyMySQL==1.0.3
pymysql-pool==0.3.7
pytz==2021.3
requests==2.31.0
tweepy==4.14.0